requirements.txt
pysimplegui==4.60.5.0
huggingface_hub==0.27.1
python-dotenv==1.0.1
pydantic==2.10.5